Move state helper functions to separate module
