+++ CartPole/cartpole_model.py
@@ -1,6 +1,6 @@
 from types import SimpleNamespace
 from typing import Union
-from CartPole._CartPole_mathematical_helpers import create_cartpole_state, cartpole_state_varname_to_index
+from CartPole.state_utilities import create_cartpole_state, cartpole_state_varname_to_index
 
 import numpy as np
