+++ Controllers/controller_custom_mpc_scipy.py
@@ -4,7 +4,7 @@ import scipy.optimize
 
 from CartPole.cartpole_model import P_GLOBALS
 
-from CartPole._CartPole_mathematical_helpers import create_cartpole_state, cartpole_state_varname_to_index, \
+from CartPole.state_utilities import create_cartpole_state, cartpole_state_varname_to_index, \
     cartpole_state_indices_to_varnames
 from Modeling.TF.TF_Functions.Network import *
 from Predictores.predictor_ideal import predictor_ideal
