+++ Controllers/controller_do_mpc.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from Controllers.template_controller import template_controller
 from CartPole.cartpole_model import P_GLOBALS, Q2u, cartpole_ode_namespace
-from CartPole._CartPole_mathematical_helpers import cartpole_state_vector_to_namespace
+from CartPole.state_utilities import cartpole_state_vector_to_namespace
 
 from types import SimpleNamespace
