+++ Controllers/controller_lqr.py
@@ -7,7 +7,7 @@ import scipy
 import numpy as np
 
 from Controllers.template_controller import template_controller
-from CartPole._CartPole_mathematical_helpers import create_cartpole_state, cartpole_state_varname_to_index
+from CartPole.state_utilities import create_cartpole_state, cartpole_state_varname_to_index
 from CartPole.cartpole_model import cartpole_jacobian, u_max, s0
 
 class controller_lqr(template_controller):
