+++ Controllers/controller_rnn_as_mpc_tf.py
@@ -1,7 +1,7 @@
 from Modeling.TF.TF_Functions.Network import *
 
 from Controllers.template_controller import template_controller
-from CartPole._CartPole_mathematical_helpers import create_cartpole_state, cartpole_state_varname_to_index
+from CartPole.state_utilities import create_cartpole_state, cartpole_state_varname_to_index
 
 
 RNN_FULL_NAME = 'GRU-5IN-64H1-64H2-1OUT-0'
