+++ Modeling/TF/TF_Functions/Dataset.py
@@ -160,7 +160,7 @@ class Dataset(keras.utils.Sequence):
 
 
 
-from CartPole._CartPole_mathematical_helpers import create_cartpole_state, cartpole_state_varname_to_index
+from CartPole.state_utilities import create_cartpole_state, cartpole_state_varname_to_index
 from Predictores.predictor_ideal import predictor_ideal
 class DatasetRandom(keras.utils.Sequence):
     def __init__(self,
