+++ run_data_generator.py
@@ -1,6 +1,6 @@
 from CartPole import CartPole
 from CartPole.cartpole_model import s0, TrackHalfLength, P_GLOBALS
-from CartPole._CartPole_mathematical_helpers import cartpole_state_varname_to_index
+from CartPole.state_utilities import cartpole_state_varname_to_index
 
 from time import sleep
 import timeit
@@ -11,17 +11,17 @@ import numpy as np
 # csv = '2500/Test'  # Name with which data is saved, consecutive experiments will be save with increasing index attached
 # csv = '25/Train/Train'
 # csv = '25/Test/Test'
-csv = '25/Validate/Validate'
-number_of_experiments = 10  # How many experiments will be generated
-save_mode = 'online'  # It was intended to save memory usage, but it doesn't seems to help. Leave it false.
+csv = 'Test'
+number_of_experiments = 1  # How many experiments will be generated
+save_mode = 'offline'  # It was intended to save memory usage, but it doesn't seems to help. Leave it false.
 
 # Timescales
 dt_simulation_DataGen = 0.002  # simulation timestep
-dt_controller_update_DataGen = 0.1
+dt_controller_update_DataGen = 0.02
 dt_save_DataGen = 0.1
 
 # CartPole settings - check the effect first in GUI before you launch big data generation
-length_of_experiment_DataGen = 600.0  # Length of each experiment in s
+length_of_experiment_DataGen = 10.0  # Length of each experiment in s
 controller_DataGen = 'do-mpc'  # Controller which should be used in generated experiment
 # Possible options for controller:
 # 'manual-stabilization', 'do-mpc', 'lqr'
